declare main view function in blueprint.py
